+++ ms-pago/src/main/resources/db/migration/V1__crear_tabla_pago.sql
@@ -1,10 +1,10 @@
 CREATE TABLE pago (
-                       id_pago INT AUTO_INCREMENT PRIMARY KEY,
-                       id_ticket BIGINT NOT NULL,
-                       monto_total DOUBLE NOT NULL,
-                       metodo_pago VARCHAR(50) NOT NULL,
-                       estado VARCHAR(50) NOT NULL,
-                       fecha_pago DATETIME NOT NULL
+                      id_pago BIGINT AUTO_INCREMENT PRIMARY KEY,
+                      id_ticket BIGINT NOT NULL,
+                      monto_total DOUBLE NOT NULL,
+                      metodo_pago VARCHAR(50) NOT NULL,
+                      estado VARCHAR(50) NOT NULL,
+                      fecha_pago DATETIME NOT NULL
 );
 
 INSERT INTO pago (id_ticket, monto_total, metodo_pago, estado, fecha_pago)
